+++ doc/versionHistory.rst
@@ -6,6 +6,14 @@
 Version History
 ##################
 
+.. _lsst.ts.imsim-0.5.1:
+
+-------------
+0.5.1
+-------------
+
+* Add convertZernikesToPsfWidth utility.
+
 .. _lsst.ts.imsim-0.5.0:
 
 -------------